OpenCL: Associate cl::context to single device
For some unknown reason we are creating cl::contexts with multiple
devices even tough in CUDA (and presumably in HIP) contexts are
associated to exactly one device.

This is a problem on 2-tile PVC where creating multi-device contexts
cause unnecessary memory synchronization. This causes, for example,
HeCBench/hybridsort-hip to slow down by 1.66x